+++ word-prediction--kb-bert/.bumpversion.cfg
@@ -12,6 +12,3 @@ replace = version = "{new_version}"
 search = __version__ = "{current_version}"
 replace = __version__ = "{new_version}"
 
-[bumpversion:file:tests/test_version.py]
-search = assert word_prediction.__version__ == "{current_version}"
-replace = assert word_prediction.__version__ == "{new_version}"
