--- word-prediction--kb-bert/tests/test_version.py
@@ -1,5 +1,0 @@
-import word_prediction
-
-
-def test_version() -> None:
-    assert word_prediction.__version__ == "0.3.0"
